[feat] create tha bus
